+++ DungeonBasic Scripts/Monster/Monster 0 Pattern.cs.meta
