+++ DungeonBasic Scripts/Monster/Monster 1 Pattern.cs.meta
