+++ DungeonBasic Scripts/Monster/PatternManager.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public abstract class PatternManager : MonoBehaviour
+{
+    public abstract void Battle(Animator anim, MonsterData data);
+}
