+++ DungeonBasic Scripts/Monster/PatternManager.cs.meta
